Multiple small fixes
FSE : can handle table size > 12 bits
FSE : can select ILP or simple mode
Data : removed data files that can be generated using ProbaGen
Data : added win98-lz4-run, compressed by FSE
Data : added benchmark result readme file

+++ test/probaGenerator.c
@@ -87,7 +87,7 @@ static void generate(void* buffer, size_t buffSize, double p)
     char* op = (char*) buffer;
     char* oend = op + buffSize;
 
-    DISPLAY("\nGenerating %i MB with P=%.2f%%\n", buffSize >> 20, p*100);
+    DISPLAY("\nGenerating %i MB with P=%.2f%%\n", (int)(buffSize >> 20), p*100);
 
     // Build Table
     while (remaining)
